backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "app-tracker"
version = "0.0.1"
description = "A Django-based web application tracker with SvelteKit frontend"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: Django",
]
dependencies = [
    "Django>=4.2.0",
    "djangorestframework>=3.14.0",
    "python-decouple>=3.8",
    "mysqlclient>=2.2.0",
    "django-cors-headers>=4.0.0",
    "gunicorn>=21.0.0",
    "whitenoise>=6.5.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/app-tracker"
Repository = "https://github.com/yourusername/app-tracker.git"

[project.scripts]
app-tracker = "app_tracker.management.commands.runserver:main"

[tool.setuptools]
packages = ["app_tracker"]
package-dir = {"" = "."}

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
